new route paths, client and admin views
